requirements.txt
joblib==1.4.2
matplotlib==3.10.0
numpy==2.2.2
pandas==2.2.3
plotly==5.24.1
PyYAML==6.0.2
scikit_learn==1.6.1
seaborn==0.13.2
shap==0.46.0
statsmodels==0.14.4
streamlit==1.40.2
